+++ app/components/files/file-state.js
@@ -11,9 +11,11 @@ import { promiseWhen } from '../helpers/sugar';
 class FileState extends RoutedState {
     @observable currentFile = null;
     @observable currentFolder = null;
+    @observable previewFile = null;
     localFileMap = observable.map();
     store = fileStore;
     _prefix = 'files';
+    selectedFile = null;
 
     @action async init() {
         this.currentFolder = fileStore.folders.root;
@@ -90,6 +92,7 @@ class FileState extends RoutedState {
     }
 
     @action resetSelection() {
+        this.selectedFile = null;
         this.selected.forEach(f => { f.selected = false; });
     }
 
@@ -128,9 +131,7 @@ class FileState extends RoutedState {
         await promiseWhen(() => socket.authenticated);
         const chat = chatState.currentChat;
         if (!chat) throw new Error('file-state.js, uploadInline: no chat selected');
-        let renamePromise = null;
-        data.file = chat.uploadAndShareFile(data.url, data.fileName, false, () => renamePromise);
-        renamePromise = this.renamePostProcessing(data);
+        data.file = chat.uploadAndShareFile(data.url, data.fileName, false, null, data.message);
         await promiseWhen(() => data.file.fileId);
         return data.file;
     }